+++ src/Class/Enemy.cpp
@@ -4,7 +4,7 @@ void Enemy::moveEnemy(float dt) {
     coordinateX += static_cast<int>(direct) * (this->getSpeed()) * dt;
 }
 
-int Enemy::getCoordinateXOfEnemy() {
+float Enemy::getCoordinateXOfEnemy() {
     return coordinateX;
 }
 
@@ -12,7 +12,7 @@ bool Enemy::checkCollision(int leftHitbox, int rightHitbox) {
     return coordinateX + this->getWidth() / 2 > leftHitbox && coordinateX - this->getWidth() / 2 < rightHitbox;
 }
 
-Enemy::Enemy(Direction direct, int coordinateXOfEnemy)
+Enemy::Enemy(Direction direct, float coordinateXOfEnemy)
 : direct(direct)
 , coordinateX(coordinateXOfEnemy) {
 }
